requirements.txt
# EPUB App System Requirements

# Web Framework
fastapi>=0.115.13
uvicorn[standard]>=0.30.0
jinja2>=3.1.4
python-multipart>=0.0.12

# Text Processing
ebooklib>=0.19
beautifulsoup4>=4.12.0
PyYAML>=6.0

# MCP Support
fastmcp>=2.9.0

# Embedding and Search (MLX + FAISS)
numpy>=1.24.0
faiss-cpu>=1.7.4
mlx>=0.18.0
mlx-lm>=0.20.0
types-PyYAML>=6.0.12

# Utilities
requests>=2.32.4
aiofiles>=24.0.0

# Type Hints
types-requests>=2.32.4.20250611
